+++ src/util/getTodayDate.ts
@@ -4,5 +4,9 @@ export function getTodayDate() {
   const month = String(today.getMonth() + 1).padStart(2, '0')
   const day = String(today.getDate()).padStart(2, '0')
 
-  return `${year}년 ${month}월 ${day}일, 지혜에서만 만나볼 수 있는 초특가 상품들`
+  return {
+    year,
+    month,
+    day,
+  }
 }
